Add files via upload
You will find three datasets, which are from the City and County Data Book, composed of 1994 and 1990 demographic data on all counties in the U.S. along with presidential voting tendencies for the 1992 U.S. presidential election. These datasets consist of 16 variables. In this project, an R Shiny app was developed to conduct and present statistical analysis on demographic and political data across U.S. states.